third_party/WebKit/LayoutTests/inspector/sources/debugger/script-formatter-breakpoints-4-expected.txt

Tests the script formatting is working fine with breakpoints.


Running: testBreakpointSetInOriginalAndRemovedInFormatted
Adding breakpoint.
Formatting.
Removing breakpoint.
Unformatting.
